🔨 Refactor: 자체 회원가입/로그인 쿠키 전달 수정
